+++ src/vo/EvaluationVO.java
@@ -2,12 +2,12 @@ package vo;
 
 public class EvaluationVO {
 	//	评分
-	public int score;
+	public double score;
 	//	评价
 	public String comment;
 	
 	
-	public EvaluationVO(int score, String comment) {
+	public EvaluationVO(double score, String comment) {
 		this.score = score;
 		this.comment = comment;
 	}
